server/requirements.txt
runpod 
supabase
pytorch
numpy
gdown

# sam2
#git+https://github.com/facebookresearch/sam2.git

# pytorch3d
git+https://github.com/facebookresearch/pytorch3d.git